refactor: MatchResponse 폴더 변경, Record는 VO가 아님
