+++ badminton-api/src/main/java/org/badminton/api/match/model/dto/SinglesMatchResponse.java
@@ -1,4 +1,4 @@
-package org.badminton.api.match.model.vo;
+package org.badminton.api.match.model.dto;
 
 public record SinglesMatchResponse(
 	Long participant1Id,
